+++ Client/Properties/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyProduct("Pathoschild.Http.Client")]
-[assembly: AssemblyCopyright("Copyright 2012–2017")]
-[assembly: AssemblyVersion("3.0.0.0")]
-[assembly: AssemblyFileVersion("3.0.0.0")]