+++ Client/Retry/RetryConfig.cs
@@ -52,5 +52,15 @@ namespace Pathoschild.Http.Client.Retry
         {
             return this.GetDelayCallback(retry, response);
         }
+
+        /// <summary>Get a retry config indicating no request should be retried.</summary>
+        public static IRetryConfig None()
+        {
+            return new RetryConfig(
+                maxRetries: 0,
+                shouldRetry: response => false,
+                getDelay: (attempt, response) => TimeSpan.Zero
+            );
+        }
     }
 }