+++ Tests/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: Guid("0b18c453-907b-4e39-901e-b5c068908b7b")]
-[assembly: AssemblyProduct("Pathoschild.Http.Client")]
-[assembly: AssemblyCopyright("Copyright 2012–2017")]
-[assembly: AssemblyVersion("3.0.0.0")]
-[assembly: AssemblyFileVersion("3.0.0.0")]